+++ src/components/ImageLoader/BlurhashCanvas.jsx
@@ -1,7 +1,19 @@
 import { useRef, useState, useEffect } from 'react';
 import { decode } from 'blurhash';
 
-export default ({ style, hash, punch, ratio, width, height, blurhashRef }) => {
+const BLANK = 'data:image/gif;base64,R0lGODlhAQABAAD/ACwAAAAAAQABAAACADs=';
+
+export default ({
+  style,
+  hash,
+  punch,
+  ratio,
+  width,
+  height,
+  imgClass,
+  imgStyle,
+  blurhashRef,
+}) => {
   if (!blurhashRef) {
     blurhashRef = useRef();
   }
@@ -9,17 +21,19 @@ export default ({ style, hash, punch, ratio, width, height, blurhashRef }) => {
 
   useEffect(() => {
     const canvas = blurhashRef.current;
-    if (canvas) {
+    if (canvas && styleHeight) {
       const pixels = decode(hash, width, height, punch);
       const ctx = canvas.getContext('2d');
       const imageData = ctx.createImageData(width, height);
       imageData.data.set(pixels);
       ctx.putImageData(imageData, 0, 0);
     }
-  }, [hash, width, height, punch, blurhashRef]);
+  }, [hash, width, height, punch, blurhashRef, styleHeight]);
 
   const aspectRatio = blurhashRef.current?.style.aspectRatio;
   useEffect(() => {
+    // require('./fast-blurhash');
+    // return;
     const canvas = blurhashRef.current;
     if (canvas) {
       if (canvas?.style.aspectRatio) {
@@ -34,11 +48,31 @@ export default ({ style, hash, punch, ratio, width, height, blurhashRef }) => {
     }
   }, [ratio, blurhashRef, aspectRatio]);
 
-  return (
+  // We only create a canvas after we have processed the original image's
+  // computed style. Until then, we render a blank image to make sure
+  // that it gets the same dimensions as the original image. Making it an
+  // image lets us mimic the original image's computed css style.
+  return styleHeight ? (
     <canvas
       style={{ ...style, height: styleHeight }}
       height={height}
       width={width}
+      ref={blurhashRef}
+    />
+  ) : (
+    <img
+      src={BLANK}
+      alt=""
+      className={imgClass ? imgClass + ' blurhash' : 'blurhash'}
+      style={imgStyle}
+      data={JSON.stringify({
+        hash,
+        punch,
+        ratio,
+        width,
+        height,
+        canvasStyle: style,
+      })}
       ref={blurhashRef}
     />
   );